+++ components/interactive/InteractiveAGiXT.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { ReactNode, useMemo } from 'react';
-import Chat from '@/components/conversation/Chat';
+import Chat from '@/components/conversation/conversation';
 import { Overrides } from '@/components/interactive/InteractiveConfigContext';
 
 export type FormProps = {
